% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_endpoint.R
\name{api_fields}
\alias{api_fields}
\title{Parse DHIS2 endpoint fields}
\usage{
api_fields(...)
}
\arguments{
\item{...}{DHIS2 query parameters.}
}
\value{
A character string.
}
\description{
This function parses DHIS2 endpoint fields into a character string.
}
\examples{
# select all the fields
api_fields("*")
# select only name, id, shortName and code
api_fields("name", "id", "shortName", "code")
# select fields in a collection
api_fields("name", organisationUnits = c("name", "id", "code"))
}
