% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_endpoint.R
\name{api_filter}
\alias{api_filter}
\title{Parse DHIS2 filter}
\usage{
api_filter(x, filter, obj)
}
\arguments{
\item{x}{An endpoint field.}

\item{filter}{A DHIS2 filter object.}

\item{obj}{Objects to filter.}
}
\description{
Short hand for specifying an filter parameter in an \code{\link{api_query}}.
}
\examples{
api_filter("name", "eq", "TZ FP A360 - Conversion rate < 20")
api_filter("name", "ilike", "TZ FP A360 - Conversion rate < 20")
}
