% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2_api.R
\name{api_fields}
\alias{api_fields}
\title{Parse DHIS2 API fields}
\usage{
api_fields(...)
}
\description{
Parse DHIS2 API fields
}
\examples{
# select all the fields
api_fields("*")
# select only name, id, shortName and code
api_fields("name","id","shortName","code")
# select fields in collection
api_fields("name", organisationUnits = c("name","id","code"))
}
