% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_order.R
\name{api_order}
\alias{api_order}
\title{Parse DHIS2 order}
\usage{
api_order(x, by = c("asc", "iasc", "desc", "idesc"))
}
\arguments{
\item{x}{An endpoint field.}

\item{by}{Type of ordering. Use `asc` to order in ascending order, `desc` in
descending order. The variants `iasc` and `idesc` are case insensitive.}
}
\description{
Short hand for specifying an order parameter in an \code{\link{api_query}}.
}
\examples{
# order name in ascending order
api_order("name", by = "asc")
# order name in descending order
api_order("name", by = "desc")
# example in a query to order by ascending order
api_query(order = api_order("name", "asc"))
# order data elements in descending order
api_endpoint(resource = "dataElements", order = api_order("name", "desc"))
}
