% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{api_version}
\alias{api_version}
\title{API version}
\usage{
api_version(version = 29)
}
\arguments{
\item{version}{An integer. The API version.}
}
\value{
Integer
}
\description{
Sets an API version. Last three versions of DHIS2 are supported.
}
\keyword{internal}
