% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{basic_auth}
\alias{basic_auth}
\title{Authenticate a User Account to DHIS2}
\usage{
basic_auth(url = NULL, username = NULL, password = NULL)
}
\arguments{
\item{url}{Server url, character string.}

\item{username}{DHIS2 username or Account name, character string.}

\item{password}{Password of the DHIS2 account, character string.}
}
\value{
An S3 object with the current user information.
}
\description{
Authenticate an account to a DHIS2 instance.
}
\details{
\code{dhis_auth} supplies an account credentials to a DHIS2 instance for
authentication. It returns an S3 object with current information of the
account.
}
