% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2_api.R
\name{create_api_link}
\alias{create_api_link}
\alias{parse_api_args}
\title{Create a DHIS2 web API link}
\usage{
create_api_link(
  resources = "resources",
  id = NULL,
  ...,
  fields = api_fields(...),
  filter = NULL,
  version = api_version(),
  skip_paging = TRUE,
  page = 1L,
  page_size = 50L,
  order = NULL,
  order_by = c("asc", "iasc", "desc", "idesc")
)

parse_api_args(
  resources = "resources",
  id = NULL,
  ...,
  version = api_version(),
  fields = NULL,
  filter = NU,
  skip_paging = "FALSE",
  page = 1L,
  page_size = 50L,
  order = NULL,
  order_by = c("asc", "iasc", "desc", "idesc")
)
}
\arguments{
\item{resources}{A character. The name of the resources to browse, such as
dataElements.}

\item{id}{A character. An identifier of the resource or the identifiable object.}

\item{...}{Additional params to parsed before the query option.}

\item{version}{An integer. The web API version.}

\item{skip_paging}{Logical. Indicates whether to return lists of elements in
pages. The default option is FALSE.}

\item{page}{An integer. Defines which page number to return. The default
option is 1.}

\item{page_size}{An integer. Defines the number of elements to return for
each page. The default size in 50.}

\item{order}{A character. Order the output using a specified order, only
properties that are both persisted and simple (no collections, idObjects
etc) are supported. iasc and idesc are case insensitive sorting.}
}
\value{
A string. DHIS2 web API link

An S3 object of class dhis2_api_args
}
\description{
This function creates an API link for browsing the DHIS2 web
resources from R. Use these utilities to construct, define, and parse
API parameters quickly.

\itemize{
 \item `api_fields` returns a string of the API fields
 \item `api_filter` returns a string of the API filters
 \item `api_version` returns a number, a version of the web API. By default, this is sets this to version `29`
 \item `api_order` returns a string of the API order params.
}

This function parse the API params in a list.
}
\details{
\code{create_api_link} provides an interface to construct, define or parse
DHIS2 web API parameters easily. By default, it returns the endpoint
`api/29/resources`, with paging set to `TRUE` and displays the first page
with the default size of `fifty`.
}
\examples{
# the resources endpoint
create_api_link()
# navigate to dataElements endpoint
create_api_link("dataElements")
# select data elements name, id and shortNames
create_api_link("dataElements", fields = api_fields("name","id","shortNames"))
# arrange the data elements name in ascending order
create_api_link("dataElements", fields = api_fields("name","id","shortNames"), order = "name", order_by = "asc")
}
