% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dhis2_error_message}
\alias{dhis2_error_message}
\title{Provide additional error information}
\usage{
dhis2_error_message(resp)
}
\arguments{
\item{resp}{response to be captured}
}
\value{
An error character vector
}
\description{
Provide additional error information
}
