% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2r-package.R
\docType{package}
\name{dhis2r-package}
\alias{dhis2r}
\alias{dhis2r-package}
\title{dhis2r: Client for the 'DHIS2' Web API}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Connect and pull data from a 'DHIS2 (District Health Information Software 2)' instance into R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amanyiraho/dhis2r}
  \item Report bugs at \url{https://github.com/amanyiraho/dhis2r/issues}
}

}
\author{
\strong{Maintainer}: Robinson Amanyiraho \email{amanyiraho@gmail.com} (\href{https://orcid.org/0000-0002-7747-1367}{ORCID}) [copyright holder]

}
\keyword{internal}
