% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2r.R
\name{dhis2r_query}
\alias{dhis2r_query}
\title{Define HTTP request}
\usage{
dhis2r_query(http_method, url, ..., params = NULL)
}
\arguments{
\item{url}{URL.}

\item{...}{Additional}
}
\description{
Define an http request.
}
