% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_dhis2.R
\name{login_dhis2}
\alias{login_dhis2}
\title{Login to a DHIS2 instance}
\usage{
login_dhis2(host)
}
\arguments{
\item{host}{A string. Server domain name}
}
\value{
Login details
}
\description{
Login to a DHIS2 instance
}
\examples{
base <- "clone.psi-mis.org"
login_dhis2(base)
}
