% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoint.R
\name{modify_api_endpoint}
\alias{modify_api_endpoint}
\title{Modify a DHIS2 API endpoint link}
\usage{
modify_api_endpoint(
  endpoint,
  ...,
  id = NULL,
  ref = NULL,
  ref_id = NULL,
  version = NULL
)
}
\arguments{
\item{endpoint}{A DHIS2 endpoint, example `dataElements`, `dataElementGroups`, e.tc.}

\item{...}{Query parameters parsed to \code{\link{api_query}}.}

\item{id}{The unique identifier for an object in the endpoint.}

\item{ref}{A reference within the endpoint.}

\item{ref_id}{The unique identifier of a referenced object in the endpoint.}

\item{version}{An integer, the version of DHIS2 web API. The default version
is `current supported version`, you can override this using the
\code{\link{api_version}} or by specifying the value explicitly.}
}
\value{
A string, the DHIS2 web API link.
}
\description{
Short hand to modify a `DHIS2 like` API endpoint link from R. This is useful
if you need to define or parse API parameters quickly.
}
\examples{
# default api link
modify_api_endpoint()
# navigate to dataElements endpoint
modify_api_endpoint(endpoint = "dataElements")
# select data elements name, id and shortNames
modify_api_endpoint(endpoint = "dataElements", fields = c("name", "id", "shortNames"))
# arrange the data elements name in ascending order
modify_api_endpoint(
  endpoint = "dataElements", fields = c("name", "id", "shortNames"),
  order = api_order("name", by = "asc")
)
}
