% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_endpoint.R
\name{modify_endpoint}
\alias{modify_endpoint}
\title{Modify a DHIS2 endpoint}
\usage{
modify_endpoint(
  endpoint = NULL,
  id = NULL,
  ref = NULL,
  ref_id = NULL,
  version = NULL,
  ...
)
}
\arguments{
\item{endpoint}{A DHIS2 resource or on an API endpoint link, for example
`dataElements` or  `api/dataElements`.}

\item{id}{The unique identifier of an object in the endpoint, a character
string of size 11.}

\item{ref}{An endpoint reference, a character string.}

\item{ref_id}{The unique identifier of an object in the referenced endpoint, a
character string of size 11.}

\item{version}{A version of the web API.}

\item{...}{Additional parameters passed to the \code{\link{api_query}}.}
}
\value{
A URL, a DHIS2 web API link.
}
\description{
Short hand to modify a `DHIS2 like` resource endpoint or an API link from R.
This is useful when defining or modifying a DHIS2 API link.
}
\details{
\code{modify_endpoint} defines or modifies an API link from an endpoint or API
params. It parses the API parameters first into a list before building the API
link.
}
\examples{
modify_endpoint() # default api link
modify_endpoint(endpoint = "dataElements") # point to dataElements endpoint
# refer to a specific dataElement object
modify_endpoint(endpoint = "dataElements", id = "ID")
# modify an object ID of an API link, `api/dataElements/ID`
modify_endpoint(endpoint = "api/dataElements/ID", id = "ID2")
# modify the version of an API link, `api/dataElements/ID`
version <- 29L
modify_endpoint(endpoint = "api/dataElements/ID", version = version)
# select data elements name, id and shortNames
modify_endpoint(endpoint = "dataElements", fields = c("name", "id", "shortNames"))
# arrange the data elements name in ascending order
modify_endpoint(
  endpoint = "dataElements", fields = c("name", "id", "shortNames"),
  order = api_order("name", by = "asc")
)
}
