% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2_api.R
\name{parse_api_args}
\alias{parse_api_args}
\alias{build_api_link}
\title{Parse DHIS2 API link/endpoint params in a list}
\usage{
parse_api_args(
  endpoint = "resources",
  id = NULL,
  ...,
  version = api_version(),
  fields = NULL,
  filter = NU,
  skip_paging = "FALSE",
  page = 1L,
  page_size = 50L,
  order = NULL,
  order_by = c("asc", "iasc", "desc", "idesc")
)

build_api_link(api_args)
}
\arguments{
\item{endpoint}{A character string, the endpoint to browse from, should be a
valid DHIS2 resource for example; `dataElements`.}

\item{id}{A character string, an identifier of the DHIS2 resource.}

\item{...}{Additional parameters to reference the endpoint.}

\item{version}{An integer, the version of the web API. The default version is
`v29`, you can override this using the \code{\link{api_version}} or by
setting the value explicitly.}

\item{fields}{A character string, the resource fields to extract for example;
you can select `name`, `id` and the `shortName` from a `dataElement`
resource. Use \code{\link{api_fields}} to parse the resource fields more
consistently.}

\item{filter}{A character string, defines the resource filter parameters. Use
\code{\link{api_filter}} to parse the resource filters consistently.}

\item{skip_paging}{Logical, indicates whether to return lists of elements in
pages. The default option is FALSE.}

\item{page}{An integer, defines which page number to return. The default
option is page 1.}

\item{page_size}{An integer, defines the number of elements to return for
each page. The default size in 50.}

\item{order}{A character string, a resource field to arrange/sort.}

\item{order_by}{A character string, the specified order, only properties that
are both persisted and simple (no collections, idObjects etc) are
supported. iasc and idesc are case insensitive sorting.}

\item{api_args}{A list. DHIS2 web API parameters}
}
\value{
An S3 object of class dhis2_api_args.
}
\description{
This function parse the API link/endpoint params in a list.
}
\details{
By default, \code{parse_api_args}, returns an S3 object from the
  `api/29/resources` endpoint with paging set to `TRUE` and displays the
  first page with the default size of `fifty`. You can override or just
  specify these options directly.
}
\note{
\code{\link{api_fields}} returns a character string of the API fields.

\code{\link{api_filter}} returns a character string of the API filters.

\code{\link{api_version}} returns an integer, the version of the web
  API. By default, this is sets this to `v29`.

\code{\link{api_order}} returns a character string, the order params of
  the API.
}
\examples{
api_args <- parse_api_args()
build_api_link(api_args)
}
