% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_endpoint.R
\name{parse_api_params}
\alias{parse_api_params}
\title{Parse DHIS2 API endpoint parameters}
\usage{
parse_api_params(
  endpoint,
  ...,
  id = NULL,
  ref = NULL,
  ref_id = NULL,
  version = NULL
)
}
\arguments{
\item{endpoint}{A DHIS2 endpoint, example `dataElements`, `dataElementGroups`, e.tc.}

\item{...}{Query parameters parsed to \code{\link{api_query}}.}

\item{id}{The unique identifier for an object in the endpoint.}

\item{ref}{A reference within the endpoint.}

\item{ref_id}{The unique identifier of a referenced object in the endpoint.}

\item{version}{An integer, the version of DHIS2 web API. The default version
is `current supported version`, you can override this using the
\code{\link{api_version}} or by specifying the value explicitly.}
}
\value{
An S3 object of class `api_params`.
}
\description{
This function parse DHIS2 API endpoint parameters into a list object.
}
\examples{
# default endpoint endpoint
api_params <- parse_api_params()
# api params with query object
api_params <- parse_api_params(endpoint = "dataElements", field = c("name", "code", "id"))
}
