% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_endpoint.R
\name{parse_endpoint}
\alias{parse_endpoint}
\title{Parse DHIS2 API endpoint parameters}
\usage{
parse_endpoint(
  endpoint = NULL,
  id = NULL,
  ref = NULL,
  ref_id = NULL,
  version = NULL,
  ...
)
}
\arguments{
\item{endpoint}{A DHIS2 resource or on an API endpoint link, for example
`dataElements` or  `api/dataElements`.}

\item{id}{The unique identifier of an object in the endpoint, a character
string of size 11.}

\item{ref}{An endpoint reference, a character string.}

\item{ref_id}{The unique identifier of an object in the referenced endpoint, a
character string of size 11.}

\item{version}{A version of the web API.}

\item{...}{Additional parameters passed to the \code{\link{api_query}}.}
}
\value{
An S3 object of class `api_params`.
}
\description{
This function parse DHIS2 API endpoint parameters into a list object.
}
\examples{
# default endpoint endpoint
endpoint_params <- parse_endpoint()
endpoint_params
# endpoint params with a DHIS2 query
endpoint_params <- parse_endpoint(endpoint = "dataElements", field = c("name", "code", "id"))
endpoint_params
}
