% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{relative_periods}
\alias{relative_periods}
\title{Relative periods in DHIS2}
\format{
A data frame with 30 rows and 2 columns:
\describe{
  \item{period_type}{Period type}
  \item{period}{Relative period to be used}
}
}
\source{
"https://docs.dhis2.org/en/develop/using-the-api/dhis-core-version-239/introduction.html"
}
\usage{
data(relative_periods)
}
\description{
The relative periods are relative to the current date and allow e.g. for creating dynamic reports
}
\keyword{datasets}
