% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reponse_check.R
\name{status_check}
\alias{status_check}
\title{Evaluate a server response}
\usage{
status_check(req)
}
\arguments{
\item{req}{An HTTP request}
}
\value{
an error message or an invisible object.
}
\description{
\code{status_check} evaluates the status  code of an HTTP request and returns an appropriate message.
}
