% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{win_secret}
\alias{win_secret}
\title{Retrieve secret from windows}
\usage{
win_secret(cred_label)
}
\arguments{
\item{cred_label}{A string. Usually the name of your secret file.}
}
\value{
An invisible string.
}
\description{
\code{win_secret} returns an invisible decrypted password stored with
Micrososft data protection API.
}
\details{
Invisible means that the password won't be displayed in the console, but can
be assigned to a variable or used inline.'Under the hood, \code{win_secret} is
using \code{\link[keyringr]decrypt_dpapi_pw} to help retrieve and unlock
passwords stored with the Microsoft data protection API. It Requires
Powershell to be installed and execution policy set to RemoteSigned.
}
\seealso{
\code{\link[keyringr]{decrypt_dpapi_pw}}
}
