% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_fields.R
\name{api_fields}
\alias{api_fields}
\title{Parse DHIS2 API fields}
\usage{
api_fields(...)
}
\description{
This function parse API fields params in standard DHIS2 format.
}
\examples{
# select all the fields
api_fields("*")
# select only name, id, shortName and code
api_fields("name","id","shortName","code")
# select fields in a collection
api_fields("name", organisationUnits = c("name","id","code"))
}
