% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dhis2r-package.R
\docType{package}
\name{dhis2r-package}
\alias{dhis2r}
\alias{dhis2r-package}
\title{dhis2r: An R client for the Dhis2 web API}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Pull data from a DHIS2 instance into R without the hassle of logging into the DHIS2 user interface every day, week, month, quarter, or year to download the same data or make the same routine data analysis steps based on fixed or relative periods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/amanyiraho/dhis2r}
  \item \url{http://www.amanyiraho.com/dhis2r/}
  \item Report bugs at \url{https://github.com/amanyiraho/dhis2r/issues}
}

}
\author{
\strong{Maintainer}: Robinson Amanyiraho \email{amanyiraho@gmail.com} (\href{https://orcid.org/0000-0002-7747-1367}{ORCID})

Other contributors:
\itemize{
  \item AFENET [copyright holder, funder]
}

}
\keyword{internal}
