% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_link.R
\name{parse_api_params}
\alias{parse_api_params}
\title{Parse DHIS2 API link/endpoint params in a list}
\usage{
parse_api_params(resources = NULL, id = NULL, path = NULL, version = NULL, ...)
}
\arguments{
\item{id}{A character string, an identifier of the DHIS2 resource.}

\item{path}{A character string, specifies a sub directory of the resource.}

\item{version}{An integer, the version of the web API. The default version is
`v29`, you can override this using the \code{\link{api_version}} or by
setting the value explicitly.}

\item{...}{Addional parameters parsed to \code{\link{api_query}}}

\item{resource}{A character string, the endpoint to browse from, should be a
valid DHIS2 resource for example; `dataElements`.}
}
\value{
An S3 object of class `dhis2_api_params`
}
\description{
This function returns an API link/endpoint for browsing DHIS2 web
resources from R. Useful for constructing, defining, or parsing API
parameters quickly.
}
\examples{
# default resource endpoint
api_params <- parse_api_params()
# api params with query object
api_params <- parse_api_params("dataElements", field = c("name","code","id"))
}
